requirements.txt
pandas
pyarrow
boto3
requests
python-dotenv
